requirements.txt
# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.7.0
cryptography>=38.0.0
pycryptodome>=3.15.0

# Blockchain SDK
hfc>=0.9.0
grpcio>=1.50.0
protobuf>=4.21.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Visualization (optional)
matplotlib>=3.5.0

# Utilities
python-dotenv>=0.21.0
pyyaml>=6.0
